Refactor TestimonialsPage component by removing unused metadata and enhancing the layout with new components for improved user experience.
