+++ app/testimonials/metadata.ts
@@ -1,0 +1,21 @@
+import type { Metadata } from "next"
+
+export const metadata: Metadata = {
+  title: "Client Reviews – Noida's Top-Rated Construction Company",
+  description: "Read real client reviews and testimonials. Discover why gharbuildr.com is rated 4.8 stars on Google and is Noida's preferred construction partner.",
+  openGraph: {
+    title: "Client Reviews – Noida's Top-Rated Construction Company",
+    description: "Read real client reviews and testimonials. Discover why gharbuildr.com is rated 4.8 stars on Google and is Noida's preferred construction partner.",
+    images: [
+      {
+        url: "/greater-noida-modern-villa.png",
+        width: 1200,
+        height: 630,
+        alt: "gharbuildr.com - Client Testimonials",
+      },
+    ],
+  },
+  alternates: {
+    canonical: "https://gharbuildr.com/testimonials",
+  },
+}
